+++ src/main/java/de/guntram/mcmod/durabilityviewer/DurabilityViewer.java
@@ -1,13 +1,14 @@
 package de.guntram.mcmod.durabilityviewer;
 
 import de.guntram.mcmod.durabilityviewer.client.gui.GuiItemDurability;
-import de.guntram.mcmod.durabilityviewer.handler.ConfigurationHandler;
-import de.guntram.mcmod.fabrictools.ConfigurationProvider;
+import de.guntram.mcmod.durabilityviewer.config.Configs;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.minecraft.client.option.KeyBinding;
 import net.minecraft.client.util.InputUtil;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import static org.lwjgl.glfw.GLFW.GLFW_KEY_H;
 
@@ -16,17 +17,16 @@ public class DurabilityViewer implements ClientModInitializer {
     public static final String MODNAME = "Durability Viewer";
 
     public static DurabilityViewer instance;
-    private static ConfigurationHandler confHandler;
     private static String changedWindowTitle;
     private KeyBinding showHide;
+    public static final Logger LOGGER = LogManager.getLogger("DurabilityViewer");
 
     @Override
     public void onInitializeClient() {
         setKeyBindings();
-        confHandler = ConfigurationHandler.getInstance();
-        ConfigurationProvider.register(MODNAME, confHandler);
-        confHandler.load(ConfigurationProvider.getSuggestedFile(MODID));
         changedWindowTitle = null;
+
+        Configs.loadFromFile();
     }
 
     public static void setWindowTitle(String s) {