+++ src/main/java/de/guntram/mcmod/durabilityviewer/config/WarnMode.java
@@ -1,34 +1,20 @@
-package de.guntram.mcmod.durabilityviewer.client.gui;
+package de.guntram.mcmod.durabilityviewer.config;
 
+import com.google.common.collect.ImmutableList;
+import fi.dy.masa.malilib.config.IConfigOptionList;
 import fi.dy.masa.malilib.config.IConfigOptionListEntry;
 import fi.dy.masa.malilib.util.StringUtils;
 
-public enum Corner implements IConfigOptionListEntry {
-    BOTTOM_RIGHT("Bottom-Right", "durabilityviewer.config.bottom_right"),
-    BOTTOM_LEFT("Bottom-Left", "durabilityviewer.config.bottom_left"),
-    TOP_RIGHT("Top-Right", "durabilityviewer.config.top_right"),
-    TOP_LEFT("Top-Left", "durabilityviewer.config.top_left");
-
-    public boolean isLeft() {
-        return this == TOP_LEFT || this == BOTTOM_LEFT;
-    }
-
-    public boolean isRight() {
-        return this == TOP_RIGHT || this == BOTTOM_RIGHT;
-    }
-
-    public boolean isTop() {
-        return this == TOP_LEFT || this == TOP_RIGHT;
-    }
-
-    public boolean isBottom() {
-        return this == BOTTOM_LEFT || this == BOTTOM_RIGHT;
-    }
+public enum WarnMode implements IConfigOptionListEntry {
+    NONE("None", "durabilityviewer.config.warnmode.none"),
+    SOUND("Sound", "durabilityviewer.config.warnmode.sound"),
+    VISUAL("Visual", "durabilityviewer.config.warnmode.visual"),
+    BOTH("Both", "durabilityviewer.config.warnmode.both");
 
     private final String configString;
     private final String translationKey;
 
-    private Corner(String configString, String translationKey) {
+    private WarnMode(String configString, String translationKey) {
         this.configString = configString;
         this.translationKey = translationKey;
     }
@@ -69,14 +55,14 @@ public enum Corner implements IConfigOptionListEntry {
     }
 
     @Override
-    public Corner fromString(String name)
+    public WarnMode fromString(String name)
     {
         return fromStringStatic(name);
     }
 
-    public static Corner fromStringStatic(String name)
+    public static WarnMode fromStringStatic(String name)
     {
-        for (Corner mode : Corner.values())
+        for (WarnMode mode : WarnMode.values())
         {
             if (mode.configString.equalsIgnoreCase(name))
             {
@@ -84,6 +70,6 @@ public enum Corner implements IConfigOptionListEntry {
             }
         }
 
-        return Corner.BOTTOM_RIGHT;
+        return WarnMode.SOUND;
     }
 }