+++ src/main/java/de/guntram/mcmod/durabilityviewer/mixin/WindowTitleMixin.java
@@ -6,7 +6,7 @@
 package de.guntram.mcmod.durabilityviewer.mixin;
 
 import de.guntram.mcmod.durabilityviewer.DurabilityViewer;
-import de.guntram.mcmod.durabilityviewer.handler.ConfigurationHandler;
+import de.guntram.mcmod.durabilityviewer.config.Configs;
 import net.minecraft.client.MinecraftClient;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -20,7 +20,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class WindowTitleMixin {
     @Inject(method = "getWindowTitle", at = @At("HEAD"), cancellable = true)
     private void patchWindowTitle(CallbackInfoReturnable<String> cir) {
-        if (ConfigurationHandler.showPlayerServerName()) {
+        if (Configs.Settings.SetWindowTitle.getBooleanValue()) {
             if (DurabilityViewer.getWindowTitle() != null) {
                 cir.setReturnValue(DurabilityViewer.getWindowTitle());
             }